restore ability to expire without having history database

+++ news/expire.c
@@ -12,12 +12,20 @@
  *
  * Author:  Kai Uwe Rommel <rommel@ars.muc.de>
  * Created: Sun Aug 15 1993
+ *
+ *    $Id$
+ *
+ *    $Log$
+ *
  */
 
-static char *rcsid = "$Id: EXPIRE.C 1.1 1993/09/05 10:56:49 rommel Exp $";
-static char *rcsrev = "$Revision: 1.1 $";
+static char *rcsid = "$Id: expire.c 1.2 1993/10/31 21:32:55 ahd Exp ahd $";
+static char *rcsrev = "$Revision: 1.2 $";
 
-/* $Log: EXPIRE.C $
+/* $Log: expire.c $
+ * Revision 1.2  1993/10/31  21:32:55  ahd
+ * Delete unreferenced variables
+ *
  * Revision 1.1  1993/09/05  10:56:49  rommel
  * Initial revision
  * */
@@ -67,7 +75,21 @@ extern struct grp *group_list;   /* List of all groups */
 /*                        Internal prototypes                         */
 /*--------------------------------------------------------------------*/
 
-static void ExpireAll( char **groups, const time_t expire_date );
+static void HistoryExpireAll( char **groups, const time_t expire_date );
+
+static void ExpireAll( const time_t expire_date );
+
+static void  ExpireGroup( const char *group,
+                      const time_t expire_date );
+
+static void ExpireOneGroup( struct grp *cur_grp,
+                      const time_t expire_date );
+
+static void ExpireDirectory( struct grp *cur_grp,
+                      const time_t expire_date,
+                      const char *directory );
+
+static boolean numeric( char *start);
 
 static void usage( void );
 
@@ -139,7 +161,7 @@ void main( int argc, char **argv)
    }
 
    if (optind != argc)
-     groups = argv + optind;
+      groups = argv + optind;
 
 /*--------------------------------------------------------------------*/
 /*                             Initialize                             */
@@ -167,17 +189,26 @@ void main( int argc, char **argv)
 /*                       Load the active file                         */
 /*--------------------------------------------------------------------*/
 
-   history = open_history("history");
-   new_history = open_history("newhist");
+   if ( bflag[F_HISTORY] )
+   {
+      history = open_history("history");
+      new_history = open_history("newhist");
+   }
 
    get_active();              /* Get sequence numbers for groups from
                                  active file                      */
 
-   /* think as if we would purge everything at the moment :-) */
-   for ( cur_grp = group_list; cur_grp != NULL;
-         cur_grp = cur_grp->grp_next )
-     if (stricmp(cur_grp->grp_name, "junk") != 0)
-       cur_grp->grp_low = cur_grp->grp_high;
+/*--------------------------------------------------------------------*/
+/*                  Chain together groups to process                  */
+/*--------------------------------------------------------------------*/
+
+   if ( bflag[F_HISTORY] )
+   {
+      for ( cur_grp = group_list; cur_grp != NULL;
+            cur_grp = cur_grp->grp_next )
+         if (stricmp(cur_grp->grp_name, "junk") != 0)
+            cur_grp->grp_low = cur_grp->grp_high;
+   } /* if */
 
 /*--------------------------------------------------------------------*/
 /*                  Compute times for expiring files                  */
@@ -194,7 +225,15 @@ void main( int argc, char **argv)
 /*    history database                                                */
 /*--------------------------------------------------------------------*/
 
-   ExpireAll(groups, expire_date);
+   if ( bflag[F_HISTORY] )
+      HistoryExpireAll(groups, expire_date );
+   else if ( groups != NULL )
+   {
+      while( argc-- > optind )
+         ExpireGroup( *groups++, expire_date );
+   }
+   else
+      ExpireAll( expire_date );
 
 /*--------------------------------------------------------------------*/
 /*                         Clean up and exit                          */
@@ -202,22 +241,25 @@ void main( int argc, char **argv)
 
    put_active();
 
-   close_history(history);
-   close_history(new_history);
+   if ( bflag[F_HISTORY] )
+   {
+      close_history(history);
+      close_history(new_history);
 
-   mkfilename(file_old, E_newsdir, "oldhist.dir");
-   mkfilename(file_new, E_newsdir, "history.dir");
-   unlink(file_old);
-   rename(file_new, file_old);
-   mkfilename(file_old, E_newsdir, "newhist.dir");
-   rename(file_old, file_new);
+      mkfilename(file_old, E_newsdir, "oldhist.dir");
+      mkfilename(file_new, E_newsdir, "history.dir");
+      unlink(file_old);
+      rename(file_new, file_old);
+      mkfilename(file_old, E_newsdir, "newhist.dir");
+      rename(file_old, file_new);
 
-   mkfilename(file_old, E_newsdir, "oldhist.pag");
-   mkfilename(file_new, E_newsdir, "history.pag");
-   unlink(file_old);
-   rename(file_new, file_old);
-   mkfilename(file_old, E_newsdir, "newhist.pag");
-   rename(file_old, file_new);
+      mkfilename(file_old, E_newsdir, "oldhist.pag");
+      mkfilename(file_new, E_newsdir, "history.pag");
+      unlink(file_old);
+      rename(file_new, file_old);
+      mkfilename(file_old, E_newsdir, "newhist.pag");
+      rename(file_old, file_new);
+   }
 
    if ( total_articles_purged)
       printmsg(1,"Purged %ld articles, %ld cross postings (%ld bytes).",
@@ -281,12 +323,12 @@ static void SetGroupLower(char *histentry)
 }
 
 /*--------------------------------------------------------------------*/
-/*    E x p i r e A l l                                               */
+/*    H i s t o r y E x p i r e A l l                                 */
 /*                                                                    */
 /*    Expire all requested news groups                                */
 /*--------------------------------------------------------------------*/
 
-static void ExpireAll( char **groups, const time_t expire_date )
+static void HistoryExpireAll( char **groups, const time_t expire_date )
 {
    char *messageID, *histentry;
    time_t article_date;
@@ -338,7 +380,230 @@ static void ExpireAll( char **groups, const time_t expire_date )
      total_bytes_kept      += remaining * article_size;
    }
 
+} /* HistoryExpireAll */
+
+/*--------------------------------------------------------------------*/
+/*    E x p i r e A l l                                               */
+/*                                                                    */
+/*    Expire all defined news groups                                  */
+/*--------------------------------------------------------------------*/
+
+static void ExpireAll( const time_t expire_date )
+{
+   struct grp *cur_grp = group_list;
+
+   while ( cur_grp != NULL )
+   {
+      ExpireOneGroup( cur_grp, expire_date  );
+                                    /* Clean up this group           */
+
+      cur_grp = cur_grp->grp_next;  /* Then clean up the next group  */
+   }
 } /* Expire_All */
+
+/*--------------------------------------------------------------------*/
+/*    E x p i r e G r o u p                                           */
+/*                                                                    */
+/*    Clean up one group by name                                      */
+/*--------------------------------------------------------------------*/
+
+static void  ExpireGroup( const char *group,
+                      const time_t expire_date )
+{
+   struct grp *cur_grp = group_list;
+   struct grp *target = NULL;
+
+/*--------------------------------------------------------------------*/
+/*         Search the list of groups for the requested group          */
+/*--------------------------------------------------------------------*/
+
+   while ( (cur_grp != NULL) && (target == NULL))
+   {
+      if ( equal( cur_grp->grp_name, group ))
+         target = cur_grp;
+
+      cur_grp = cur_grp->grp_next;  /* Then clean up the next group  */
+   }
+
+/*--------------------------------------------------------------------*/
+/*   If we found the group, process it, otherwise report the error    */
+/*--------------------------------------------------------------------*/
+
+   if ( target == NULL )
+      printmsg(0,"Unable to locate active group %s", group );
+   else
+      ExpireOneGroup( target, expire_date );
+                                    /* Clean up this group           */
+
+} /* ExpireGroup */
+
+/*--------------------------------------------------------------------*/
+/*    E x p i r e O n e G r o u p                                     */
+/*                                                                    */
+/*    Clean up one group by name                                      */
+/*--------------------------------------------------------------------*/
+
+static void ExpireOneGroup( struct grp *cur_grp,
+                      const time_t expire_date )
+{
+   char groupdir[FILENAME_MAX];
+
+   printmsg(3,"Processing news group %s", cur_grp->grp_name );
+
+/*--------------------------------------------------------------------*/
+/*                     Set up the directory names                     */
+/*--------------------------------------------------------------------*/
+
+   ImportNewsGroup( groupdir, cur_grp->grp_name, 0 );
+
+/*--------------------------------------------------------------------*/
+/*                      Process the directories                       */
+/*--------------------------------------------------------------------*/
+
+   ExpireDirectory( cur_grp, expire_date, groupdir );
+
+} /* ExpireOneGroup */
+
+/*--------------------------------------------------------------------*/
+/*    E x p i r e D i r e c t o r y                                   */
+/*                                                                    */
+/*    Clean up one group by name                                      */
+/*--------------------------------------------------------------------*/
+
+static void ExpireDirectory( struct grp *cur_grp,
+                      const time_t expire,
+                      const char *directory )
+{
+   int articles_purged   = 0;/* Count of files actually deleted        */
+   int articles_kept     = 0;/* Count of files actually deleted        */
+   long bytes_purged = 0;    /* Bytes freed on disk from deletions     */
+   long bytes_kept   = 0;    /* Bytes left on disk total               */
+
+   long low = LONG_MAX;  /* Oldest article number left             */
+
+   DIR *dirp;
+   struct direct *dp;
+
+/*--------------------------------------------------------------------*/
+/*                Open up the directory for processing                */
+/*--------------------------------------------------------------------*/
+
+   if ((dirp = opendirx(directory,"*.*")) == nil(DIR))
+   {
+      printmsg(3, "ExpireDirectory: couldn't opendir() %s", directory);
+      cur_grp->grp_low = cur_grp->grp_high;
+      return;
+   } /* if */
+
+/*--------------------------------------------------------------------*/
+/*                 Switch to directory for processing                 */
+/*--------------------------------------------------------------------*/
+
+   CHDIR( directory );
+
+/*--------------------------------------------------------------------*/
+/*              Look for the next file in the directory               */
+/*--------------------------------------------------------------------*/
+
+   while((dp = readdir(dirp)) != nil(struct direct))
+   {
+
+/*--------------------------------------------------------------------*/
+/*                      Archive/expire this file?                     */
+/*--------------------------------------------------------------------*/
+
+      if ( numeric( dp->d_name ))/* Article format name?             */
+      {                          /* Yes --> Examine it closer        */
+
+         printmsg(6,"Processing file %s from %s",
+                  dp->d_name, dater( dp->d_modified, NULL));
+
+         if ( dp->d_modified < expire )   /* Long in the tooth?      */
+         {                       /* Yes --> Move it on out           */
+
+            printmsg( 4,"Purging file %s from %s", dp->d_name, directory );
+            unlink( dp->d_name );
+            articles_purged++;
+            bytes_purged += dp->d_size;
+
+         } /* if ( dp->d_modified < expire ) */
+
+/*--------------------------------------------------------------------*/
+/*    If the article is valid and still in the main news              */
+/*    directory, determine if it is the lowest article left           */
+/*--------------------------------------------------------------------*/
+
+         else {
+            long article = 0;
+            char *digit = dp->d_name;
+
+            while( *digit )
+               article = article * 10 + (*digit++ - '0');
+
+            low = min( article, low );
+
+            bytes_kept += dp->d_size;
+            articles_kept ++;
+
+         } /* else if ( archive != NULL ) */
+      } /* if ( numeric( dp->d_name ) */
+
+   } /* while */
+
+/*--------------------------------------------------------------------*/
+/*            Update lowest article available to the users            */
+/*--------------------------------------------------------------------*/
+
+   if ( low == LONG_MAX )
+       cur_grp->grp_low = cur_grp->grp_high;
+   else
+       cur_grp->grp_low = low;
+
+/*--------------------------------------------------------------------*/
+/*           Close up the directory and report what we did            */
+/*--------------------------------------------------------------------*/
+
+   closedir(dirp);
+
+   if ( articles_purged )
+      printmsg(2,"%s: Purged %d articles (%ld bytes),"
+                    " left alone %d articles (%ld bytes).",
+                  cur_grp->grp_name,
+                  articles_purged, bytes_purged,
+                  articles_kept, bytes_kept );
+   else if ( articles_kept )
+      printmsg(2,"%s: Left alone %d articles (%ld bytes).",
+                  cur_grp->grp_name,
+                  articles_kept, bytes_kept );
+
+   total_articles_kept     += articles_kept;
+   total_articles_purged   += articles_purged;
+   total_bytes_kept        += bytes_kept;
+   total_bytes_purged      += bytes_purged;
+
+} /* ExpireDirectory */
+
+/*--------------------------------------------------------------------*/
+/*    n u m e r i c                                                   */
+/*                                                                    */
+/*    Examines string, returns true if numeric with period            */
+/*--------------------------------------------------------------------*/
+
+ static boolean numeric( char *start)
+ {
+   char *number = start;
+
+   while (*number != '\0')
+   {
+      if (!isdigit(*number) && (*number != '.'))
+         return FALSE;
+
+      number++;
+   }
+
+   return TRUE;
+ } /* numeric */
+
 
 /*--------------------------------------------------------------------*/
 /*    u s a g e                                                       */
